requirements.txt
asgiref==3.7.2
blinker==1.6.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.9.0
contourpy==1.0.7
cycler==0.11.0
deprecation==2.1.0
Django==4.2.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
docopt==0.6.2
Flask==2.3.3
fonttools==4.39.4
func-timeout==4.3.5
functions-framework==3.4.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-compute==1.15.0
google-cloud-secret-manager==2.17.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2


MarkupSafe==2.1.3

numpy==1.24.3
packaging==23.1
pipreqs==0.4.13
plumbum==1.8.2
proto-plus==1.23.0
protobuf
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.0.9
requests==2.31.0
rpyc==5.3.1
rsa==4.9



six==1.16.0


tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
watchdog==3.0.0
Werkzeug==2.3.7
yarg==0.1.9
